Convert `describeSpecialist` to TypeScript.
