+++ packages/agent-kit/agent.kit.json
@@ -11836,7 +11836,7 @@
               "id": "runJavascript-0",
               "type": "runJavascript",
               "configuration": {
-                "code": "const run = function describeSpecialist(inputs) {\n  const { $inputSchema, $outputSchema, persona, task } = inputs;\n\n  const inputSchema = {\n    type: \"object\",\n    properties: {\n      ...$inputSchema.properties,\n      in: {\n        title: \"Context in\",\n        description: \"Incoming conversation context\",\n        type: \"array\",\n        items: {\n          type: \"object\",\n          behavior: [\"llm-content\"],\n        },\n        examples: [],\n      },\n      task: {\n        title: \"Task\",\n        description:\n          \"(Optional) Provide a specific task with clear instructions for the worker to complete using the conversation context. Use mustache-style {{params}} to add parameters.\",\n        type: \"object\",\n        default: '{\"role\":\"user\",\"parts\":[{\"text\":\"\"}]}',\n        behavior: [\"llm-content\", \"config\"],\n        examples: [],\n      },\n      persona: {\n        type: \"object\",\n        behavior: [\"llm-content\", \"config\"],\n        title: \"Persona\",\n        description:\n          \"Describe the worker's skills, capabilities, mindset, and thinking process. Use mustache-style {{params}} to add parameters.\",\n        default: '{\"role\":\"user\",\"parts\":[{\"text\":\"\"}]}',\n        examples: [],\n      },\n    },\n    required: [],\n  };\n\n  const params = unique([\n    ...collectParams(textFromLLMContent(persona)),\n    ...collectParams(textFromLLMContent(task)),\n  ]);\n\n  const props = Object.fromEntries(\n    params.map((param) => [\n      toId(param),\n      {\n        title: toTitle(param),\n        description: `The value to substitute for the parameter \"${param}\"`,\n        type: \"string\",\n      },\n    ])\n  );\n\n  const required = params.map(toId);\n\n  return mergeSchemas(inputSchema, $outputSchema, props);\n\n  function mergeSchemas(inputSchema, outputSchema, properties) {\n    return {\n      inputSchema: {\n        ...inputSchema,\n        properties: {\n          ...inputSchema.properties,\n          ...properties,\n        },\n        required: [...(inputSchema.required || []), ...required],\n      },\n      outputSchema: outputSchema,\n    };\n  }\n\n  function toId(param) {\n    return `p-${param}`;\n  }\n\n  function toTitle(id) {\n    const spaced = id?.replace(/[_-]/g, \" \");\n    return (\n      (spaced?.at(0)?.toUpperCase() ?? \"\") +\n      (spaced?.slice(1)?.toLowerCase() ?? \"\")\n    );\n  }\n\n  function textFromLLMContent(content) {\n    return content?.parts.map((item) => item.text).join(\"\\n\") || \"\";\n  }\n\n  function unique(params) {\n    return Array.from(new Set(params));\n  }\n\n  function collectParams(text) {\n    if (!text) return [];\n    const matches = text.matchAll(/{{(?<name>[\\w-]+)}}/g);\n    return Array.from(matches).map((match) => match.groups?.name || \"\");\n  }\n};",
+                "code": "const run = function describeSpecialist(inputs){const{$inputSchema,$outputSchema,persona,task}=inputs;const inputSchema={type:\"object\",properties:{...$inputSchema.properties,in:{title:\"Context in\",description:\"Incoming conversation context\",type:\"array\",items:{type:\"object\",behavior:[\"llm-content\"]},examples:[]},task:{title:\"Task\",description:\"(Optional) Provide a specific task with clear instructions for the worker to complete using the conversation context. Use mustache-style {{params}} to add parameters.\",type:\"object\",default:'{\"role\":\"user\",\"parts\":[{\"text\":\"\"}]}',behavior:[\"llm-content\",\"config\"],examples:[]},persona:{type:\"object\",behavior:[\"llm-content\",\"config\"],title:\"Persona\",description:\"Describe the worker's skills, capabilities, mindset, and thinking process. Use mustache-style {{params}} to add parameters.\",default:'{\"role\":\"user\",\"parts\":[{\"text\":\"\"}]}',examples:[]}},required:[]};const params=unique([...collectParams(textFromLLMContent(persona)),...collectParams(textFromLLMContent(task))]);const props=Object.fromEntries(params.map(param=>[toId(param),{title:toTitle(param),description:`The value to substitute for the parameter \"${param}\"`,type:\"string\"}]));const required=params.map(toId);return mergeSchemas(inputSchema,$outputSchema,props);function mergeSchemas(inputSchema2,outputSchema,properties){return{inputSchema:{...inputSchema2,properties:{...inputSchema2.properties,...properties},required:[...inputSchema2.required||[],...required]},outputSchema}}__name(mergeSchemas,\"mergeSchemas\");function toId(param){return`p-${param}`}__name(toId,\"toId\");function toTitle(id){const spaced=id?.replace(/[_-]/g,\" \");return(spaced?.at(0)?.toUpperCase()??\"\")+(spaced?.slice(1)?.toLowerCase()??\"\")}__name(toTitle,\"toTitle\");function textFromLLMContent(content){return content?.parts.map(item=>{return\"text\"in item?item.text:\"\"}).join(\"\\n\")||\"\"}__name(textFromLLMContent,\"textFromLLMContent\");function unique(params2){return Array.from(new Set(params2))}__name(unique,\"unique\");function collectParams(text){if(!text)return[];const matches=text.matchAll(/{{(?<name>[\\w-]+)}}/g);return Array.from(matches).map(match=>match.groups?.name||\"\")}__name(collectParams,\"collectParams\")};",
                 "inputSchema": {
                   "type": "object",
                   "properties": {},
