+++ packages/agent-kit/src/future/templating.js
@@ -4,110 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { describeSpecialist, content, describeContent };
-
-/**
- * The describer for the "Specialist" v2 component.
- */
-function describeSpecialist(inputs) {
-  const { $inputSchema, $outputSchema, persona, task } = inputs;
-
-  const inputSchema = {
-    type: "object",
-    properties: {
-      ...$inputSchema.properties,
-      in: {
-        title: "Context in",
-        description: "Incoming conversation context",
-        type: "array",
-        items: {
-          type: "object",
-          behavior: ["llm-content"],
-        },
-        examples: [],
-      },
-      task: {
-        title: "Task",
-        description:
-          "(Optional) Provide a specific task with clear instructions for the worker to complete using the conversation context. Use mustache-style {{params}} to add parameters.",
-        type: "object",
-        default: '{"role":"user","parts":[{"text":""}]}',
-        behavior: ["llm-content", "config"],
-        examples: [],
-      },
-      persona: {
-        type: "object",
-        behavior: ["llm-content", "config"],
-        title: "Persona",
-        description:
-          "Describe the worker's skills, capabilities, mindset, and thinking process. Use mustache-style {{params}} to add parameters.",
-        default: '{"role":"user","parts":[{"text":""}]}',
-        examples: [],
-      },
-    },
-    required: [],
-  };
-
-  const params = unique([
-    ...collectParams(textFromLLMContent(persona)),
-    ...collectParams(textFromLLMContent(task)),
-  ]);
-
-  const props = Object.fromEntries(
-    params.map((param) => [
-      toId(param),
-      {
-        title: toTitle(param),
-        description: `The value to substitute for the parameter "${param}"`,
-        type: "string",
-      },
-    ])
-  );
-
-  const required = params.map(toId);
-
-  return mergeSchemas(inputSchema, $outputSchema, props);
-
-  function mergeSchemas(inputSchema, outputSchema, properties) {
-    return {
-      inputSchema: {
-        ...inputSchema,
-        properties: {
-          ...inputSchema.properties,
-          ...properties,
-        },
-        required: [...(inputSchema.required || []), ...required],
-      },
-      outputSchema: outputSchema,
-    };
-  }
-
-  function toId(param) {
-    return `p-${param}`;
-  }
-
-  function toTitle(id) {
-    const spaced = id?.replace(/[_-]/g, " ");
-    return (
-      (spaced?.at(0)?.toUpperCase() ?? "") +
-      (spaced?.slice(1)?.toLowerCase() ?? "")
-    );
-  }
-
-  function textFromLLMContent(content) {
-    return content?.parts.map((item) => item.text).join("\n") || "";
-  }
-
-  function unique(params) {
-    return Array.from(new Set(params));
-  }
-
-  function collectParams(text) {
-    if (!text) return [];
-    const matches = text.matchAll(/{{(?<name>[\w-]+)}}/g);
-    return Array.from(matches).map((match) => match.groups?.name || "");
-  }
-}
+export { content, describeContent };
 
 /**
  * The guts of the "Content" component.
