+++ packages/agent-kit/tests/templating.ts
@@ -5,18 +5,15 @@
  */
 
 import { describe, it } from "node:test";
-import {
-  substitute,
-  content,
-  describeSpecialist,
-  describeContent,
-} from "../src/future/templating.js";
-import { InlineDataCapabilityPart, LLMContent } from "@google-labs/breadboard";
+import { content, describeContent } from "../src/future/templating.js";
+import { InlineDataCapabilityPart } from "@google-labs/breadboard";
 import { deepStrictEqual } from "node:assert";
+import { describeSpecialist, substitute } from "../src/templating.js";
+import { LlmContent } from "../src/context.js";
 
 describe("Templating", () => {
   it("does simple substitution", async () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const persona = llmContent("user", "You're a {{character}}, {{name}}.");
     const task = llmContent("user", "Act like a {{character}}.");
     const character = "wizard";
@@ -35,7 +32,7 @@ describe("Templating", () => {
   });
 
   it("does simple LLM Content object substitution", async () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const persona = llmContent("user", "You're a {{character}}, {{name}}.");
     const task = llmContent("user", "Act like a {{character}}.");
     const character = llmContent("user", "wizard");
@@ -55,7 +52,7 @@ describe("Templating", () => {
   });
 
   it("does part-splicing LLM Content object substitution", async () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const persona = llmContent("user", "You're a {{character}}, {{name}}.");
     const task = llmContent("user", "Act like a {{character}}.");
     const character = llmContent("user", "wizard");
@@ -75,7 +72,7 @@ describe("Templating", () => {
   });
 
   it("works with LLM Content Array arguments", async () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const persona = llmContent("user", "You're a {{character}}, {{name}}.");
     const task = llmContent("user", "Act like a {{character}}.");
     const character = [llmContent("user", "wizard")];
@@ -95,7 +92,7 @@ describe("Templating", () => {
   });
 
   it("works with other primitive types", async () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const persona = llmContent("user", "You're a {{character}}, {{name}}.");
     const task = llmContent("user", "Act like a {{character}}.");
     const character = true;
@@ -115,7 +112,7 @@ describe("Templating", () => {
   });
 
   it("works with arbitrary objects", async () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const persona = llmContent("user", "You're a {{character}}, {{name}}.");
     const task = llmContent("user", "Act like a {{character}}.");
     const character = { wizard: true };
@@ -135,7 +132,7 @@ describe("Templating", () => {
   });
 
   it("works with other part types", async () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const persona = llmContent("user", "You're a {{character}}, {{name}}.");
     const task = llmContent("user", "Act like a {{character}}.");
     const character = llmContent("user", "wizard");
@@ -175,7 +172,7 @@ describe("Content component", () => {
   });
 
   it("correcly elides empty context", () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const template = llmContent("user", "You're a wizard, Harry.");
     const result = content({ context, template });
     deepStrictEqual(result, {
@@ -193,7 +190,7 @@ describe("Content component", () => {
   });
 
   it("does simple substitution", () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const template = llmContent("user", "Act like a {{role}}.");
     const role = "wizard";
     const result = content({ context, template, "p-role": role });
@@ -203,7 +200,7 @@ describe("Content component", () => {
   });
 
   it("does simple LLM Content object substitution", () => {
-    const context: LLMContent[] = [];
+    const context: LlmContent[] = [];
     const template = llmContent("user", "Act like a {{role}}.");
     const role = llmContent("user", "wizard");
     const result = content({ context, template, "p-role": role });
@@ -323,7 +320,7 @@ function inlineData(data: string, mimeType: string): InlineDataCapabilityPart {
 function llmContent(
   role: "user" | "model",
   ...text: (string | InlineDataCapabilityPart)[]
-): LLMContent {
+): LlmContent {
   if (text.length === 0) {
     return { parts: [{ text: "" }], role };
   }
@@ -335,5 +332,5 @@ function llmContent(
       }
       return t;
     }),
-  };
+  } as LlmContent;
 }
